More singed/unsigned work.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34204 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/api-structures.h
@@ -312,8 +312,8 @@ struct Domain_s
 	List_o_links * lol;
 	DTreeLeaf *    child;
 	Domain *       parent;
-	int            size;
-	int            start_link;  /* the link that started this domain */
+	size_t         size;
+	size_t         start_link;  /* the link that started this domain */
 	char           type;        /* one letter name */
 };
